+++ SurvivalGame/Assets/UI/BuildingUI.cs
@@ -1,7 +1,10 @@
+using System.Collections.Generic;
 using UnityEngine;
 
 public class BuildingUI : BaseScrollableCraftUI
 {
+
+    [SerializeField] protected List<GameObject> buildingsGO = new List<GameObject>();
 
     protected override void Start()
     {
@@ -15,10 +18,13 @@ public class BuildingUI : BaseScrollableCraftUI
 
     protected override void PopulateSlots()
     {
-        foreach (BuildingBlueprint buildingBlueprint in craftBlueprints)
+        foreach (GameObject building in buildingsGO)
         {
+            BuildingData data = building.GetComponent<WorldObjectBase>().GetWorldObjectData() as BuildingData;
+            BuildingBlueprint buildingBlueprint = data.blueprint;
             GameObject slot = Instantiate(slotUIPrefab, verticalGrid.transform);
-            slot.GetComponent<BuildingBlueprintSlotUI>().SetBluePrint(buildingBlueprint);
+            slot.GetComponent<BuildingBlueprintSlotUI>().SetBluePrint(buildingBlueprint, skipUpdateUI: true);
+            slot.GetComponent<BuildingBlueprintSlotUI>().SetLinkedGameObject(building);
         }
     }
 }