+++ SurvivalGame/Assets/WorldObjects/WorldObjectBase.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[DisallowMultipleComponent]
+public class WorldObjectBase : MonoBehaviour
+{
+    public virtual WorldObjectData GetWorldObjectData() => null;
+}