+++ SurvivalGame/Assets/WorldObjects/WorldObjectsData/BuildingData.cs
@@ -6,5 +6,6 @@ public class BuildingData : WorldObjectData
     [SerializeField] public int xdimension = 1;
     [SerializeField] public int zdimension = 1;
     [SerializeField] public BuildingBlueprint blueprint;
+    [SerializeField] public Sprite uiIcon;
 
 }